+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.52.1](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.0...rudder-sdk-js@2.52.1) (2025-07-30)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.16.0`
+
+### Bug Fixes
+
+* config for comscore ([#2366](https://github.com/rudderlabs/rudder-sdk-js/issues/2366)) ([f3c4567](https://github.com/rudderlabs/rudder-sdk-js/commit/f3c45674449c84524e73f5be58da26257ae79953))
+
 ## [2.52.0](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.51.4...rudder-sdk-js@2.52.0) (2025-07-21)
 
 ### Dependency Updates